Fixed formatitng issues

+++ script/HelperConfig.s.sol
@@ -10,11 +10,10 @@ abstract contract CodeConstants {
     /* VRF Mock Values */
     uint96 public constant MOCK_BASE_FEE = 0.25 ether;
     uint96 public constant MOCK_GAS_PRICE = 1e9;
-    int public constant MOCK_WEI_PER_UNIT_LINK = 4e15;
+    int256 public constant MOCK_WEI_PER_UNIT_LINK = 4e15;
     uint256 public constant ETH_SEPOLIA_CHAIN_ID = 11155111;
     uint256 public constant LOCAL_CHAIN_ID = 31337;
-    address public constant FOUNDRY_DEFAULT_SENDER =
-        0x1804c8AB1F12E6bbf3894d4083f33e07309d1f38;
+    address public constant FOUNDRY_DEFAULT_SENDER = 0x1804c8AB1F12E6bbf3894d4083f33e07309d1f38;
 }
 
 contract HelperConfig is Script, CodeConstants {
@@ -42,9 +41,7 @@ contract HelperConfig is Script, CodeConstants {
         return getConfigByChainId(block.chainid);
     }
 
-    function getConfigByChainId(
-        uint256 chainId
-    ) internal returns (NetworkConfig memory) {
+    function getConfigByChainId(uint256 chainId) internal returns (NetworkConfig memory) {
         if (networkConfigs[chainId].vrfCoordinator != address(0)) {
             return networkConfigs[chainId];
         } else if (chainId == LOCAL_CHAIN_ID) {
@@ -55,17 +52,16 @@ contract HelperConfig is Script, CodeConstants {
     }
 
     function getSepoliaETHConfig() public pure returns (NetworkConfig memory) {
-        return
-            NetworkConfig({
-                entranceFee: 0.01 ether,
-                interval: 30,
-                vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
-                gasLane: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
-                callbackGasLimit: 500000,
-                subscriptionId: 54858649297092211820447201651571043065297943636257530193770267383548256774717,
-                link: 0x779877A7B0D9E8603169DdbD7836e478b4624789,
-                account: 0xf9A17B14C32734319F27fBC910901523E4e75E3C
-            });
+        return NetworkConfig({
+            entranceFee: 0.01 ether,
+            interval: 30,
+            vrfCoordinator: 0x9DdfaCa8183c41ad55329BdeeD9F6A8d53168B1B,
+            gasLane: 0x787d74caea10b2b357790d5b5247c2f63d1d91572a9846f780606e4d953677ae,
+            callbackGasLimit: 500000,
+            subscriptionId: 54858649297092211820447201651571043065297943636257530193770267383548256774717,
+            link: 0x779877A7B0D9E8603169DdbD7836e478b4624789,
+            account: 0xf9A17B14C32734319F27fBC910901523E4e75E3C
+        });
     }
 
     function getOrCreateAnvilETHConfig() public returns (NetworkConfig memory) {
@@ -73,11 +69,8 @@ contract HelperConfig is Script, CodeConstants {
             return localNetworkConfig;
         }
         vm.startBroadcast();
-        VRFCoordinatorV2_5Mock vrfCoordinatorMock = new VRFCoordinatorV2_5Mock(
-            MOCK_BASE_FEE,
-            MOCK_GAS_PRICE,
-            MOCK_WEI_PER_UNIT_LINK
-        );
+        VRFCoordinatorV2_5Mock vrfCoordinatorMock =
+            new VRFCoordinatorV2_5Mock(MOCK_BASE_FEE, MOCK_GAS_PRICE, MOCK_WEI_PER_UNIT_LINK);
 
         LinkToken linkToken = new LinkToken();
         vm.stopBroadcast();
